Optimized string encoding
Benchmarks showed that different "methods" are faster than others. This is taken into account.
For small inputs `Convert.ToBase64String` is fastest -- probably due being intern to coreclr, so the string can be allocated and directly written to.
For a bit larger inputs, stack allocation is fastest.
The larger the input, the faster is `string.Create`.

`string.Create` has quite high overhead (valuetuple on stack, delegate invocation), so the input needs to be of reasonable size to take the
benefit of writing directly to string's memory.