ROSpan for direct mapping to static data (#65)
* Encoding-maps as ROSpan for direct static data

* ROSpan for vectors, to read from static data directly

No need to create the vectors in type initialization, then store them to the heap (once), and read from there, when a direct way is possible with C# 7.3.

* netcoreapp2.1 reverse data

For netcoreapp2.1 the vector elements were specified in reverse order (15, 14, ..., 1, 0). Now the vectors are read from static data, so ordering must be adapted -- i.e. the same ordering as netcoreapp3.0 has to be used.

* PR Feedback

* https://github.com/gfoidl/Base64/pull/65#discussion_r264725812
* https://github.com/gfoidl/Base64/pull/65#discussion_r264734952
* https://github.com/gfoidl/Base64/pull/65#discussion_r264726564